requirements.txt
pyTelegramBotAPI==4.25
python-dotenv